+++ crates/containerdebug/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Changed
+
+- All output is now wrapped in a "containerdebug" span ([#18]).
+
+[#18]: https://github.com/stackabletech/containerdebug/pull/18
+
 ## [0.1.0] - 2024-12-09
 
 ### Added